+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.0.17 / ?, 2019
 * code refactoring
+* added the blUserLocale property (two character code) to the BackendlessUser object
 
 ### 0.0.16 / July 8, 2019
 * added Codable and NSCoding support for LoadRelationsQueryBuilder